requirements.txt
-i  https://pypi.douban.com/simple
fastapi
pydantic
# mysql数据库相关
pymysql
sqlalchemy~=1.3.18
# 登陆安全相关
python-multipart
pyjwt
python-jose
passlib[bcrypt]~=1.7.2

tenacity~=6.2.0

pytest~=5.4.3
starlette~=0.13.4

